+++ src/main/java/firok/tiths/common/Traits.java
@@ -36,6 +36,7 @@ public class Traits
 	public static final AbstractTrait inky=new TraitInky(); // 墨染
 	static{inky.addItem(Items.inkySlime);}
 	public static final AbstractTrait gorgeous=new TraitGorgeous(); // 斑斓
+	public static final AbstractTrait peaceEnergetic=new TraitPeaceEnergetic(); // 平和能量
 	public static final AbstractTrait chemicalInstable=new TraitChemicalInstable(); // 化学不稳定
 	public static final AbstractTrait infernalBlazing=new TraitInfernalBlazing(); // 狱炎
 	public static final AbstractTrait dragonKiller=new TraitDragonKiller(); // 屠龙者