+++ changelog.md
@@ -2,6 +2,17 @@
 
 > All Time Stamps: GMT+8
 
+## 0.2.40.0 _2020-01-30 17:30_
+
+* 增加方块和物品 added blocks and items
+  * 托帕石矿 topaz ore block
+  * 托帕石 topaz item
+  * 托帕石块 topaz block
+* 增加材料 added materials
+  * 托帕石 topaz
+* 增加属性 added traits
+  * 平和能量 peace energetic
+
 ## 0.2.39.0 _2020-01-29 17:30_
 
 * 增加方块和物品 added blocks and items